[a/f] refactor & agent death start

+++ src/game/entities/private/player.cpp
@@ -11,11 +11,11 @@ namespace object
 
 	void Player::Update()
 	{
-		m_offset_rotation.z = m_rotation;
+		m_offset_rotation.z = (float)m_rotation;
 		float yaw = (GetTransform().GetRotation().z + m_rotation) * DEG_TO_RAD;
 
 		Vector2D direction_vec2d = { (float)cosf(yaw), (float)sin(yaw) };
-		Vector2D pos_change_vec2d = direction_vec2d * m_velocity;
+		Vector2D pos_change_vec2d = direction_vec2d * (float)m_velocity;
 
 		m_offset_pos = { pos_change_vec2d.x, pos_change_vec2d.y, 0 };
